requirement.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2.1
django-oauth-toolkit==1.2.0
django-rest-framework==0.1.0
djangorestframework==3.9.4
idna==2.8
oauthlib==3.0.1
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
